PHN-1b: four comments that overclaimed, one int that overflows, one untested member
MaskFor's summary said "8-character log mask" and returns 12 - "gvm:" plus 8 hex
characters. Corrected, and the summary now says which hash those characters come
from so it stays tied to FileNameFor.

GvMediaRegistrationTests carried three claims that were false as written. "The
first real container guard in this repository" sat next to its own note that
RotaryEncoderRegistrationTests already builds a provider and resolves from it; it
is the first guard with EAGER validation, which is the property that actually
distinguishes it. "Neither ValidateOnBuild nor ValidateScopes appears anywhere in
src/ or tests/" was present tense and false at HEAD, because that very file uses
both; it now says they appeared nowhere before this file, and says what that
costs the existing guard. And TheCacheIsASingleton_SoTheWriteLockIsProcessWide
overclaimed twice - the singleton is PROVIDER-wide, which is exactly what every
test in the file relies on, and the lock serialises writers against writers only,
since TryGetPath takes no lock at all. Renamed and reworded.

GvMediaCache's no-cache sweep computed its window as
Math.Max(60, options.MaxPlaybackSeconds * 2) in int arithmetic. That wraps
negative just past 1073741823, which would put the cutoff in the FUTURE and
delete every unprotected entry on every write. Widened to long, with the reason
recorded rather than left as a silent cast.

BuildVoicemailUri now assembles the UriBuilder inside a try and maps
UriFormatException to Transport, so no failure leaves this method carrying no
Reason. The finding named builder.Uri as the throw site; the guard covers the
Path setter too, because that is the other documented one and it sat outside the
original suggestion.

⚠ The comment on that guard says plainly that no reachable trigger was found on
.NET 10 - EscapeDataString neutralised every candidate tried, including a
200 000-character id and lone surrogates, which it folds to U+FFFD rather than
rejecting. It is kept because it makes the taxonomy hold by construction rather
than by a measurement of the current runtime, and the comment says that instead
of implying a known case exists.

GvMediaCache.RetainsEntries was public, documented and had zero callers in src/
and tests/. It is not dead - PR 3 consumes it - so it gains tests rather than a
deletion, including one pinning that it agrees with whether a write can actually
be read back. A caller that trusted it and then got a null from TryGetPath would
go back to the network during the very blackout the cache exists to survive.

+++ src/Radio.Infrastructure/External/GvMediaCache.cs
@@ -71,7 +71,11 @@ public sealed class GvMediaCache
     return string.Concat(Convert.ToHexString(hash, 0, 16).ToLowerInvariant(), ".mp3");
   }
 
-  /// <summary>The 8-character log mask for a media id. Same hash as <see cref="FileNameFor"/>.</summary>
+  /// <summary>
+  /// The 12-character log mask for a media id: the literal "gvm:" plus the first 8 hex characters
+  /// of the same SHA-256 hash <see cref="FileNameFor"/> uses, so a log line and a file on disk
+  /// correlate without either carrying the id.
+  /// </summary>
   internal static string MaskFor(string mediaId)
   {
     var hash = SHA256.HashData(Encoding.UTF8.GetBytes(mediaId));
@@ -178,7 +182,10 @@ public sealed class GvMediaCache
       }
       else
       {
-        var window = TimeSpan.FromSeconds(Math.Max(60, options.MaxPlaybackSeconds * 2));
+        // 60L and the cast are load-bearing: MaxPlaybackSeconds is an int read from configuration,
+        // and int arithmetic wraps NEGATIVE just past 1073741823, which would turn the sweep window
+        // into a cutoff in the future and delete every unprotected entry on every write.
+        var window = TimeSpan.FromSeconds(Math.Max(60L, (long)options.MaxPlaybackSeconds * 2));
         SweepOlderThan(directory, window, path, _logger);
       }
     }
